nav and style fixes
- bat pill styling
- bat pill href
- container width expanded
- small style tweaks
