+++ public/creators-landing/src/components/styled/buttons/index.js
@@ -1,18 +1,19 @@
 import styled from 'styled-components'
-import { Button, FormField } from 'grommet'
+import { Button, FormField, Box, Anchor } from 'grommet'
 
 // Large primary signup button
 export const PrimaryButton = styled(Button)`
   color: #fff;
   background: #fb542b;
   border: #fb542b;
-  height: 48px;
-  line-height: 38px;
+  font-size: 20px;
+  padding: 16px 42px;
   min-width: 160px;
   text-align: center;
   &:hover {
     background: #f43405;
     box-shadow: 0 0 0 0 !important;
+    transition: all 400ms;
   }
 `
 
@@ -24,6 +25,7 @@ export const SecondaryButton = styled(Button)`
   &:hover {
     background: rgba(255, 255, 255, 0.2);
     box-shadow: 0 0 0 0 !important;
+    transition: all 400ms;
   }
 `
 
@@ -40,5 +42,20 @@ export const StyledInput = styled(FormField)`
   & :focus-within {
     box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
     border-bottom: none !important;
+    transition: all 400ms;
   }
+`
+
+export const CardButtonContainer = styled(Box)`
+  border: 1px solid rgba(255, 255, 255, 0.4);
+  padding: 24px;
+  border-radius: 8px;
+  font-size: 18px;
+  &:hover {
+    background: rgba(255, 255, 255, 0.2);
+  }
+`
+
+export const CardButtonAnchor = styled(Anchor)`
+  flex-grow: 1;
 `
