+++ public/creators-landing/src/locale/en.js
@@ -1,6 +1,7 @@
 export default {
   nav: {
     logoHref: "/",
+    batPillHref: 'http://www.basicattentiontoken.org',
     signup: "sign up",
     signupHref: "/sign-up",
     login: "log in",
